agents/coding-bear/requirements.txt
anthropic>=0.21.0
openai>=1.12.0
pyyaml>=6.0.1
rich>=13.7.0
requests>=2.31.0
python-telegram-bot>=20.8
github3.py>=4.0.1
pygithub>=2.2.0
flask>=3.0.0
cryptography>=42.0.0